requirements.txt
langchain
langchain-community
openai
streamlit
pydub
ffmpeg-python
streamlit-flow-component==1.5.0